+++ EvergreenView/Controllers/MedicineController.cs
@@ -12,10 +12,11 @@ using System.Net.Http.Headers;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 
 namespace EvergreenView.Controllers
 {
-    [Authorize(Roles = "Admin")]
+    
     public class MedicineController : Controller
     {
         private string MedicineApiUrl = "";
@@ -45,6 +46,10 @@ namespace EvergreenView.Controllers
 
         public async Task<ActionResult> Details(int id)
         {
+            if (HttpContext.Session.GetString("r") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
             var medicine = await GetMedicineById(id);
             if (medicine == null)
                 return NotFound();
@@ -53,6 +58,10 @@ namespace EvergreenView.Controllers
 
         public async Task<ActionResult> Create()
         {
+            if (HttpContext.Session.GetString("r") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
             HttpResponseMessage responeDiseaseCategory = await client.GetAsync(MedicineCategoryApiUrl);
             string strData = await responeDiseaseCategory.Content.ReadAsStringAsync();
             var options = new JsonSerializerOptions
@@ -70,6 +79,10 @@ namespace EvergreenView.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Medicine medicine)
         {
+            if (HttpContext.Session.GetString("r") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
             string data = JsonSerializer.Serialize(medicine);
             StringContent content = new StringContent(data, Encoding.UTF8, "application/json");
             HttpResponseMessage response = client.PostAsync(MedicineApiUrl, content).Result;
@@ -92,6 +105,10 @@ namespace EvergreenView.Controllers
 
         public async Task<ActionResult> Edit(int id)
         {
+            if (HttpContext.Session.GetString("r") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
             HttpResponseMessage response = await client.GetAsync(MedicineApiUrl + "/" + id);
             string strData = await response.Content.ReadAsStringAsync();
             var options = new JsonSerializerOptions
@@ -117,6 +134,10 @@ namespace EvergreenView.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit(int id, Medicine medicine)
         {
+            if (HttpContext.Session.GetString("r") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
             string data = JsonSerializer.Serialize(medicine);
             StringContent content = new StringContent(data, Encoding.UTF8, "application/json");
             HttpResponseMessage response = client.PutAsync(MedicineApiUrl + "/" + id, content).Result;
@@ -140,6 +161,10 @@ namespace EvergreenView.Controllers
 
         public async Task<ActionResult> Delete(int id)
         {
+            if (HttpContext.Session.GetString("r") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
             var medicine = await GetMedicineById(id);
             if (medicine == null)
                 return NotFound();
@@ -151,6 +176,10 @@ namespace EvergreenView.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HttpContext.Session.GetString("r") != "Admin")
+            {
+                return RedirectToAction("Index");
+            }
             var disease = await GetMedicineById(id);
             HttpResponseMessage response = await client.DeleteAsync(MedicineApiUrl + "/" + id);
             if (response.IsSuccessStatusCode)
